put together swr state related to hackmd
